Rename go/operator to elasticjob.
