+++ docs/tutorial/tf_elasticjob_on_k8s.md
@@ -21,7 +21,7 @@ make deploy IMG=easydl/elasticjob-controller:v0.1.1
 1. Grant permission for the DLRover master to Access CRDs.
 
 ```bash
-kubectl -n dlrover apply -f dlrover/go/operator/config/rbac/default_role.yaml 
+kubectl -n dlrover apply -f dlrover/go/elasticjob/config/rbac/default_role.yaml 
 ```
 
 ## Submit an Auto-Scaling Job
