+++ go/elasticjob/api/v1alpha1/scaleplan_types.go
@@ -17,7 +17,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	commonv1 "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/common/api/v1"
+	commonv1 "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/common/api/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
