+++ go/elasticjob/pkg/controllers/elasticjob_controller.go
@@ -33,10 +33,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	elasticv1alpha1 "github.com/intelligent-machine-learning/dlrover/go/operator/api/v1alpha1"
-	common "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/common"
-	commonv1 "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/common/api/v1"
-	master "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/controllers/master"
+	elasticv1alpha1 "github.com/intelligent-machine-learning/dlrover/go/elasticjob/api/v1alpha1"
+	common "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/common"
+	commonv1 "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/common/api/v1"
+	master "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/controllers/master"
 )
 
 const (
