+++ go/elasticjob/pkg/controllers/training/task.go
@@ -21,10 +21,10 @@ import (
 	"strings"
 	"time"
 
-	elasticv1alpha1 "github.com/intelligent-machine-learning/dlrover/go/operator/api/v1alpha1"
-	common "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/common"
-	commonv1 "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/common/api/v1"
-	master "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/controllers/master"
+	elasticv1alpha1 "github.com/intelligent-machine-learning/dlrover/go/elasticjob/api/v1alpha1"
+	common "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/common"
+	commonv1 "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/common/api/v1"
+	master "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/controllers/master"
 	logger "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
