+++ go/elasticjob/pkg/controllers/training/tf_cluster.go
@@ -16,7 +16,7 @@ package training
 import (
 	"encoding/json"
 
-	commonv1 "github.com/intelligent-machine-learning/dlrover/go/operator/pkg/common/api/v1"
+	commonv1 "github.com/intelligent-machine-learning/dlrover/go/elasticjob/pkg/common/api/v1"
 	logger "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 )
